makefile done

+++ srcs/ft_printflibft.c
@@ -6,7 +6,7 @@
 /*   By: ybesbes <ybesbes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/14 12:28:50 by ybesbes           #+#    #+#             */
-/*   Updated: 2020/07/14 12:33:31 by ybesbes          ###   ########.fr       */
+/*   Updated: 2020/07/15 18:17:56 by ybesbes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,8 +29,8 @@ void	ft_putstr(char *s)
 char	*ft_strjoin(char const *s1, char const *s2)
 {
 	char	*str;
-	size_t	i;
-	size_t	j;
+	int		i;
+	int		j;
 
 	i = 0;
 	j = 0;
